working on sel testing

+++ googlemaps.py
@@ -73,7 +73,7 @@ def radius():
         flash('Please enter x and y coordinates with a radius')
         return redirect(url_for('index'))
     try:
-        if int(r) > 5:
+        if float(r) > 5:
             flash('Please enter enter a smaller radius')
             return redirect(url_for('index'))
         result = get_title_for_cords(x, y, r)
